Rename and simplify InstantExecutionProblemsListener
